mapper.* interfaces implemented Account and all necessary classes added architecture: dbSErvice -> dao -> jdbcMapper -> entityClassMetaData, entitySQLMetaData, dbExecutor, sessionManager

+++ hw18-jdbc/src/main/java/ru/otus/jdbc/DbExecutorImpl.java
@@ -25,7 +25,7 @@ public class DbExecutorImpl<T> implements DbExecutor<T> {
             pst.executeUpdate();
             try (ResultSet rs = pst.getGeneratedKeys()) {
                 rs.next();
-                return rs.getInt(1);
+                return rs.getLong(1);
             }
         } catch (SQLException ex) {
             connection.rollback(savePoint);
